# docs/swagger.yaml
basePath: /
definitions:
...
  ent.AlbumPhoto:
    properties:
      album_id:
        description: 相册ID
        type: integer
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: 相片描述
        type: string
      id:
        description: ID of the ent.
        type: integer
      image_url:
        description: 图片地址
        type: string
      name:
        description: 相片名称
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      view_count:
        description: 查看次数
        type: integer
    type: object
  ent.Category:
    properties:
      active:
        description: 是否启用
        type: boolean
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: 分类描述
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.CategoryEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CategoryQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: 分类名称
        type: string
      slug:
        description: 分类别名
        type: string
      sort_order:
        description: 排序
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  ent.CategoryEdges:
    properties:
      posts:
        description: Posts holds the value of the posts edge.
        items:
          $ref: '#/definitions/ent.Post'
        type: array
    type: object
...
  ent.Post:
    properties:
      author:
        description: 作者
        type: string
      comment_count:
        description: 评论次数
        type: integer
      content:
        description: 文章内容
        type: string
      content_type:
        allOf:
        - $ref: '#/definitions/post.ContentType'
        description: 内容类型
      copyright:
        description: 文章版权
        type: string
      cover:
        description: 文章封面
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.PostEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PostQuery when eager-loading is set.
      html_content:
        description: html文章内容
        type: string
      id:
        description: ID of the ent.
        type: integer
      is_allow_comment:
        description: 是否允许评论
        type: boolean
      is_autogen_summary:
        description: 是否自动生成摘要
        type: boolean
      is_pin_to_top:
        description: 是否置顶
        type: boolean
      is_visible:
        description: 是否可见
        type: boolean
      is_visible_after_comment:
        description: 是否评论后可见
        type: boolean
      is_visible_after_pay:
        description: 是否支付后可见
        type: boolean
      keywords:
        description: 文章关键词
        type: string
      md_content:
        description: md文章内容
        type: string
      price:
        description: 文章价格
        type: integer
      published_at:
        description: 发布时间
        type: string
      slug:
        description: 文章别名
        type: string
      status:
        allOf:
        - $ref: '#/definitions/post.Status'
        description: 状态
      summary:
        description: 文章摘要
        type: string
      title:
        description: 文章标题
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      view_count:
        description: 浏览次数
        type: integer
    type: object
  ent.PostEdges:
    properties:
      categories:
        description: Categories holds the value of the categories edge.
        items:
          $ref: '#/definitions/ent.Category'
        type: array
      tags:
        description: Tags holds the value of the tags edge.
        items:
          $ref: '#/definitions/ent.Tag'
        type: array
    type: object
...
  ent.Tag:
    properties:
      active:
        description: 是否启用
        type: boolean
      color:
        description: 标签颜色
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: 标签描述
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.TagEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TagQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: 标签名称
        type: string
      slug:
        description: 标签别名
        type: string
      sort_order:
        description: 排序
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  ent.TagEdges:
    properties:
      posts:
        description: Posts holds the value of the posts edge.
        items:
          $ref: '#/definitions/ent.Post'
        type: array
    type: object
...
  model.AlbumPhotoCreateReq:
    properties:
      album_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
    required:
    - album_id
    - image_url
    type: object
  model.AlbumPhotoUpdateReq:
    properties:
      album_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
    type: object
...
  model.CategoryCreateReq:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
    required:
    - name
    - sort_order
    type: object
  model.CategoryUpdateReq:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
    type: object
...
  model.HttpError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  model.HttpSuccess:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.PageResult-ent_AlbumPhoto:
    properties:
      records:
        items:
          $ref: '#/definitions/ent.AlbumPhoto'
        type: array
      total:
        type: integer
    type: object
  model.PageResult-ent_Category:
    properties:
      records:
        items:
          $ref: '#/definitions/ent.Category'
        type: array
      total:
        type: integer
    type: object
...
  post.ContentType:
    enum:
    - html
    - markdown
    - html
    type: string
    x-enum-varnames:
    - DefaultContentType
    - ContentTypeMarkdown
    - ContentTypeHTML
  post.Status:
    enum:
    - draft
    - draft
    - published
    - archived
    type: string
    x-enum-varnames:
    - DefaultStatus
    - StatusDraft
    - StatusPublished
    - StatusArchived
...
host: localhost:13000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /api/v1/album-photo/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的相册照片
      parameters:
      - description: 相册照片信息
        in: body
        name: album_photo
        required: true
        schema:
          $ref: '#/definitions/model.AlbumPhotoCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.AlbumPhotoCreateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建相册照片
      tags:
      - 相册照片
  /api/v1/album-photo/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定相册照片
      parameters:
      - description: 相册照片ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除相册照片
      tags:
      - 相册照片
  /api/v1/album-photo/list:
    get:
      consumes:
      - application/json
      description: 查询所有相册照片
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ent.AlbumPhoto'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询相册照片列表
      tags:
      - 相册照片
  /api/v1/album-photo/page:
    get:
      consumes:
      - application/json
      description: 查询所有相册照片分页列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-ent_AlbumPhoto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询相册照片分页列表
      tags:
      - 相册照片
  /api/v1/album-photo/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询指定相册照片的信息
      parameters:
      - description: 相册照片ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.AlbumPhoto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询相册照片
      tags:
      - 相册照片
  /api/v1/album-photo/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定相册照片的信息
      parameters:
      - description: 相册照片ID
        in: path
        name: id
        required: true
        type: string
      - description: 相册照片信息
        in: body
        name: album_photo
        required: true
        schema:
          $ref: '#/definitions/model.AlbumPhotoUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.AlbumPhotoUpdateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新相册照片
      tags:
      - 相册照片
...
  /api/v1/category/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的分类
      parameters:
      - description: 分类创建请求
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CategoryCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建分类
      tags:
      - 分类
  /api/v1/category/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定分类
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除分类
      tags:
      - 分类
  /api/v1/category/list:
    get:
      consumes:
      - application/json
      description: 查询所有分类的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ent.Category'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询分类列表
      tags:
      - 分类
  /api/v1/category/page:
    get:
      consumes:
      - application/json
      description: 查询所有分类的分页列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-ent_Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询分类分页列表
      tags:
      - 分类
  /api/v1/category/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询指定分类的信息
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询分类
      tags:
      - 分类
  /api/v1/category/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定分类的信息
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类更新请求
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CategoryUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.Category'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新分类
      tags:
      - 分类
...
swagger: "2.0"